requirements.txt
gitpython
google-generativeai
python-dotenv